Factor GridFilter settings
