+++ components/grid.js
@@ -1,10 +1,10 @@
 import { html } from "uhtml";
 import * as Props from "./props";
 import { TreeBase } from "./treebase";
+import { GridFilter } from "./gridFilter";
 import { styleString } from "./style";
 import { formatSlottedString } from "./helpers";
 import Globals from "app/globals";
-import { comparators } from "app/data";
 import "css/grid.css";
 
 /**
@@ -236,31 +236,7 @@ class Grid extends TreeBase {
   settingsDetails() {
     const props = this.propsAsProps;
     const inputs = Object.values(props).map((prop) => prop.input());
-    const filters = html`<fieldset>
-      <legend>Filters</legend>
-      <table>
-        <thead>
-          <tr>
-            <th>#</th>
-            <th>Field</th>
-            <th>Operator</th>
-            <th>Value</th>
-          </tr>
-        </thead>
-        <tbody>
-          ${this.children.map(
-            (filter, index) => html`
-              <tr id=${filter.id + "-settings"}>
-                <td>${index + 1}</td>
-                <td>${filter.field.input()}</td>
-                <td>${filter.operator.input()}</td>
-                <td>${filter.value.input()}</td>
-              </tr>
-            `,
-          )}
-        </tbody>
-      </table>
-    </fieldset>`;
+    const filters = GridFilter.FilterSettings(this.children);
     return html`<div>${filters}${inputs}</div>`;
   }
 
@@ -269,17 +245,3 @@ class Grid extends TreeBase {
   }
 }
 TreeBase.register(Grid, "Grid");
-
-export class GridFilter extends TreeBase {
-  field = new Props.Field({ hiddenLabel: true });
-  operator = new Props.Select(Object.keys(comparators), { hiddenLabel: true });
-  value = new Props.String("", { hiddenLabel: true });
-
-  /** move my parent instead of me.
-   * @param {boolean} up
-   */
-  moveUpDown(up) {
-    this.parent?.moveUpDown(up);
-  }
-}
-TreeBase.register(GridFilter, "GridFilter");
